# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "varsion": "0.0.1",
    "title": "DAS SDL API server",
    "description": "DAS SDL 서버 API swagger 문서",
    "version": "1.0.0"
  },
  "host": "localhost:3200",
  "basePath": "/das",
...
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "token",
      "description": "유저token"
    }
  },
  "paths": {
    "/user/sendEmailAuth": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "이메일 인증 코드 이메일 전송",
        "description": "사용자가 이메일 인증을 위해 인증 코드를 전송한다.",
        "content": "application/json",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login_id": { "type": "string", "example" : "altm56@gmail.com", "required" : true}
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "인증코드 발송 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/authEmailCode": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "이메일 인증 ",
        "description": "사용자가 이메일 인증을 한다",
        "content": "application/json",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login_id": { "type": "string", "example" : "altm56@gmail.com", "required" : true},
                "auth_code": { "type": "string", "example" : "krlcg1", "required" : true}
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "이메일 인증 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/signUp": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "사용자 회원가입",
        "description": "사용자가 회원가입한다.",
        "content": "application/json",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login_id": { "type": "string", "example" : "altm56@gmail.com", "required" : true},
                "login_pw": { "type": "string", "example" : "qwer123", "required" : true},
                "user_name": { "type": "string", "example" : "홍길동", "required" : true},
                "user_level" : { "type" : "number", "example" : 3, "required" : true}
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "회원가입 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/signIn": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "사용자 로그인",
        "description": "사용자가 로그인한다.",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login_id": { "type": "string", "example" : "altm56@gmail.com", "required" : true},
                "login_pw": { "type": "string", "example" : "123", "required" : true}
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "로그인 성공"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "userId" : {
                      "type" : "number",
                      "example" : 1
                    },
                    "loginId" : {
                      "type" : "string",
                      "example" : "altm56@gmail.com"
                    },
                    "isFirst" : {
                      "type" : "number",
                      "example" : 1
                    },
                    "userName" : {
                      "type" : "string",
                      "example" : "홍길동"
                    },
                    "userLevel" : {
                      "type" : "number",
                      "example" : 3
                    },
                    "userStatus" : {
                      "type" : "string",
                      "example" : "검토중"
                    },
                    "userRole" : {
                      "type" : "string",
                      "example" : "보안담당자"
                    },
                    "rgstDtm" : {
                      "type" : "string",
                      "example" : "2024/08/16 10:21:54"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/frgtEml": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "사용자 패스워드 찾기",
        "description": "사용자가 패스워드를 찾기위해 이메일을 전송한다.",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login_id": { "type": "string", "example" : "altm56@gmail.com", "required" : true}
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "비밀번호 찾기 이메일 전송 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/modify": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "사용자 비밀번호 변경",
        "description": "사용자가 비밀번호를 변경한다.",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login_id": { "type": "string", "example" : "altm56@gmail.com", "required" : true},
                "current_password" : {"type" : "string", "example" : "NbghWIoeJ(yu#408orcQ2lcQu1o#AwXcIZQOuT@G$)h!hq6GX)ujcJGzjqMuE!%5ITTnSBt5", "required" : true},
                "new_password" : { "type" : "string", "example" : "qwer123", "required" : true}
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "비밀번호 변경 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": [
              {
                "token": ""
              }
            ]
          }
        ]
      }
    },
    "/user/dvList": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "유저 리스트 출력",
        "description": "가입된 유저 리스트를 출력한다.",
        "parameters": [
          {
            "name" : "user_level",
            "description" : "계정 레벨(1: 시스템관리자, 2: 보안담당자, 3: 개발담당자)",
            "in" : "query",
            "type" : "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "개발자 리스트 출력 완료"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "userId" : {
                        "type" : "number",
                        "example" : 1
                      },
                      "loginId" : {
                        "type" : "string",
                        "example" : "altm56@gmail.com"
                      },
                      "isFirst" : {
                        "type" : "number",
                        "example" : 0
                      },
                      "userName" : {
                        "type" : "string",
                        "example" : "홍길동"
                      },
                      "userLevel" : {
                        "type" : "number",
                        "example" : 3
                      },
                      "userStatus" : {
                        "type" : "string",
                        "example" : "W"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": [
              {
                "token": ""
              }
            ]
          }
        ]
      }
    },
...
  }
}